+++ Domain/Models/T22/Solicitud.cs
@@ -11,7 +11,7 @@ namespace Domain.Models.T22
         public int IdSolicitud { get; set; }
         public int UsuarioId { get; set; }
         public string VcNombreUsuario { get; set; }
-        public int IntNumeroIdentificacionUsuario { get; set; }
+        public long IntNumeroIdentificacionUsuario { get; set; }
         public int TipoSolicitudId { get; set; }
         public string VcTipoSolicitante { get; set; }
         public int? UsuarioAsignadoId { get; set; }
